+++ authorizations/authorization-ranger/src/main/java/org/apache/gravitino/authorization/ranger/RangerHelper.java
@@ -442,64 +442,6 @@ public class RangerHelper {
             });
   }
 
-  protected RangerPolicy createPolicyAddResources(AuthorizationMetadataObject metadataObject) {
-    RangerPolicy policy = new RangerPolicy();
-    policy.setService(rangerServiceName);
-    policy.setName(metadataObject.fullName());
-    List<String> nsMetadataObject = metadataObject.names();
-    for (int i = 0; i < nsMetadataObject.size(); i++) {
-      RangerPolicy.RangerPolicyResource policyResource =
-          new RangerPolicy.RangerPolicyResource(
-              nsMetadataObject.get(i),
-              false,
-              metadataObject.type().equals(RangerMetadataObject.Type.PATH));
-      policy.getResources().put(policyResourceDefines.get(i), policyResource);
-    }
-    return policy;
-  }
-
-  protected RangerPolicy addOwnerToNewPolicy(
-      AuthorizationMetadataObject metadataObject, Owner newOwner) {
-    RangerPolicy policy = createPolicyAddResources(metadataObject);
-
-    ownerPrivileges.forEach(
-        ownerPrivilege -> {
-          // Each owner's privilege will create one RangerPolicyItemAccess in the policy
-          RangerPolicy.RangerPolicyItem policyItem = new RangerPolicy.RangerPolicyItem();
-          policyItem
-              .getAccesses()
-              .add(new RangerPolicy.RangerPolicyItemAccess(ownerPrivilege.getName()));
-          if (newOwner != null) {
-            if (newOwner.type() == Owner.Type.USER) {
-              policyItem.getUsers().add(newOwner.name());
-            } else {
-              policyItem.getGroups().add(newOwner.name());
-            }
-            // mark the policy item is created by Gravitino
-            policyItem.getRoles().add(GRAVITINO_OWNER_ROLE);
-          }
-          policy.getPolicyItems().add(policyItem);
-        });
-    return policy;
-  }
-
-  protected RangerPolicy addOwnerRoleToNewPolicy(
-      AuthorizationMetadataObject metadataObject, String ownerRoleName) {
-    RangerPolicy policy = createPolicyAddResources(metadataObject);
-
-    ownerPrivileges.forEach(
-        ownerPrivilege -> {
-          // Each owner's privilege will create one RangerPolicyItemAccess in the policy
-          RangerPolicy.RangerPolicyItem policyItem = new RangerPolicy.RangerPolicyItem();
-          policyItem
-              .getAccesses()
-              .add(new RangerPolicy.RangerPolicyItemAccess(ownerPrivilege.getName()));
-          policyItem.getRoles().add(generateGravitinoRoleName(ownerRoleName));
-          policy.getPolicyItems().add(policyItem);
-        });
-    return policy;
-  }
-
   protected void updatePolicyOwnerRole(RangerPolicy policy, String ownerRoleName) {
     // Find matching policy items based on the owner's privileges
     List<RangerPolicy.RangerPolicyItem> matchPolicyItems =
